8.0~8.15 wetube

+++ src/controllers/videoController.js
@@ -1,9 +1,11 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
   const videos = await Video.find({});
   return res.render("home", { pageTitle: "Home", videos });
 };
+
 export const search = async (req, res) => {
   const { search } = req.query;
   if (search) {
@@ -16,22 +18,28 @@ export const search = async (req, res) => {
   }
   return res.render("search", { pageTitle: "Search" });
 };
+
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
   return res.render("watch", { pageTitle: video.title, video });
 };
+
 export const getEdit = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
   }
+  if (String(video.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", { pageTitle: `Editing`, video });
 };
+
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
@@ -46,24 +54,46 @@ export const postEdit = async (req, res) => {
   });
   return res.redirect(`../${id}`);
 };
+
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== req.session.user._id) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
+
 export const getUpload = (req, res) => {
   res.render("upload", { pageTitle: " Upload Video" });
 };
+
 export const postUpload = async (req, res) => {
-  const { title, description, hashtags } = req.body;
+  const {
+    body: { title, description, hashtags },
+    file,
+    session: {
+      user: { _id },
+    },
+  } = req;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl: file.path,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
+    console.log(error);
     return res.render("upload", {
       pageTitle: " Upload Video",
       errorMessage: "error",
